cleanup, mainly moving files and ensure compatibility with python3 for print statements
