+++ .gitignore
@@ -1,1 +1,1 @@
-/config.py
+/src/config.py
